% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conditional_density.R
\name{calc_conditional_density}
\alias{calc_conditional_density}
\title{Calculate conditional density}
\usage{
calc_conditional_density(prox_matrix, data, A_col, bandwidth_grid)
}
\arguments{
\item{prox_matrix}{A proximity matrix from the random forest model.}

\item{data}{A data.frame containing the input dataset.}

\item{A_col}{A character specifying the column of interest in the dataset for density calculation.}

\item{bandwidth_grid}{A list containing bandwidth grids for each kernel function.}
}
\value{
A list of conditional density values calculated using different kernel functions.
}
\description{
This function computes the conditional density for a given proximity matrix, data, column of interest, and bandwidth grid.
}
\examples{
prox_matrix <- matrix(runif(9, 0, 1), nrow = 3)
data <- data.frame(A = c(1, 2, 3), B = c(2, 3, 4))
A_col <- "A"
bandwidth_grid <- list(gaussian_kernel = c(0.5, 1), triangular_kernel = c(0.5, 1))

calc_conditional_density(prox_matrix, data, A_col, bandwidth_grid)
}
