% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rfsrc_parallel_nll.R
\name{cv_rfsrc_parallel_nll}
\alias{cv_rfsrc_parallel_nll}
\title{Cross-Validated Random Forest SRC with Parallelization}
\usage{
cv_rfsrc_parallel_nll(
  train_data,
  outcome_col,
  k_folds = 10,
  ntree_grid,
  mtry_grid,
  nodesize_grid,
  splitrule_grid,
  kernel_grid,
  bandwidth_grid
)
}
\arguments{
\item{train_data}{A data frame containing the training data.}

\item{outcome_col}{A character string specifying the column name of the outcome variable.}

\item{k_folds}{An integer specifying the number of folds for k-fold cross-validation (default is 10).}

\item{ntree_grid}{A vector of integers specifying the grid of ntree values to search.}

\item{mtry_grid}{A vector of integers specifying the grid of mtry values to search.}

\item{nodesize_grid}{A vector of integers specifying the grid of nodesize values to search.}

\item{splitrule_grid}{A character vector specifying the grid of splitrule values to search.}

\item{kernel_grid}{A character vector specifying the grid of kernel names to search.}

\item{bandwidth_grid}{A list of numeric vectors specifying the grid of bandwidth values to search for each kernel.}
}
\value{
A list containing the best negative log-likelihood (best_nll), best ntree (best_ntree), best mtry (best_mtry),
best nodesize (best_nodesize), best splitrule (best_splitrule), best fitted model (best_model), and best kernel (best_kernel).
}
\description{
Performs a k-fold cross-validated random forest SRC with parallelization to find the best hyperparameters and kernel.
}
