% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_kernel.R
\name{dnorm_kernel}
\alias{dnorm_kernel}
\title{Dnorm kernel function}
\usage{
dnorm_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the kernel (not used in this function but kept for consistency with other kernels).}
}
\value{
The computed dnorm kernel value.
}
\description{
This function computes the dnorm kernel using a Gaussian (normal) distribution with input data and associated weights.
}
\examples{
x <- c(1, 2, 3)
data <- c(3.3, 0, 4.1, 0, 0, 2.2, 1, 4)
weights <- c(0.22, 0, 0, 0, 0, 0.11, 10.4, 0.7)
bandwidth <- 1
dnorm_kernel(x, data, weights, bandwidth)
}
