% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epanechnikov_kernel.R
\name{epanechnikov_kernel}
\alias{epanechnikov_kernel}
\title{Epanechnikov kernel}
\usage{
epanechnikov_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the Gaussian kernel.}
}
\value{
The computed Gaussian kernel value.
}
\description{
This function calculates the Epanechnikov kernel for the given input data.
}
\examples{
x <- c(1, 2, 3)
data <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3, ncol = 2)
weights <- c(0.5, 0.3, 0.2)
bandwidth <- 1

epanechnikov_kernel(x, data, weights, bandwidth)
}
