% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_kernel.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian (Radial basis function) kernel}
\usage{
gaussian_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the Gaussian kernel.}
}
\value{
The computed Gaussian kernel value.
}
\description{
This function calculates the Gaussian kernel for the given input data.
}
\examples{
x <- c(1, 2, 3)
data <- c(3.3, 0, 4.1, 0, 0, 2.2, 1, 4)
weights <- c(0.22, 0, 0, 0, 0, 0.11, 10.4, 0.7)
bandwidth <- 1
gaussian_kernel(x, data, weights, bandwidth)
}
