% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_kernel.R
\name{ks_kernel}
\alias{ks_kernel}
\title{Kernel smoothing function using the 'ks' package}
\usage{
ks_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the kernel.}
}
\value{
The computed kernel density estimation value.
}
\description{
This function calculates the kernel density estimation for the given input data using the 'ks' package.
}
\examples{
x <- c(1, 2, 3)
data <- c(1.2, 4.4, 0.11, 3.4, 9.9, 10.0, 3.4)
weights <- c(2, .4, .11, .4, .9, .0, .4)
bandwidth <- 1
ks_kernel(x, data, weights, bandwidth)
}
