% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_kernel.R
\name{triangular_kernel}
\alias{triangular_kernel}
\title{Triangular kernel}
\usage{
triangular_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the Gaussian kernel.}
}
\value{
The computed Triangular kernel value.
}
\description{
This function calculates the triangular kernel for the given input data.
}
\examples{
x <- c(1, 2, 3)
data <- c(1.2, 4.4, 0.11, 3.4, 9.9, 10.0, 3.4)
weights <- c(2, .4, .11, .4, .9, .0, .4)
bandwidth <- 1
triangular_kernel(x, data, weights, bandwidth)
}
