% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_kernel.R
\name{weighted_mean_kernel}
\alias{weighted_mean_kernel}
\title{Weighted mean kernel function}
\usage{
weighted_mean_kernel(x, data, weights, bandwidth)
}
\arguments{
\item{x}{A vector of input values to compute the kernel.}

\item{data}{A matrix of data points.}

\item{weights}{A vector of weights associated with the data points.}

\item{bandwidth}{A numeric value specifying the bandwidth of the kernel.}
}
\value{
The computed weighted mean kernel value.
}
\description{
This function computes the weighted mean kernel using a Gaussian (normal) distribution with input data and associated weights.
}
\examples{
x <- c(1, 2, 3)
data <- c(1.2, 4.4, 0.11, 3.4, 9.9, 10.0, 3.4)
weights <- c(2, .4, .11, .4, .9, .0, .4)
bandwidth <- 1
weighted_mean_kernel(x, data, weights, bandwidth)
}
